requirements.txt
pandas==2.0.3
openpyxl==3.1.2
flask==2.3.3
twilio==8.10.0
google-generativeai==0.3.2
python-dotenv==1.0.0
schedule==1.2.0
requests==2.31.0
gspread==5.10.0
oauth2client==4.1.3
